Add function input(e)
Added a function input(e) that links to the event listener associated
with the bottom half class div. Clicking anywhere on the keypad sends
the value of the target to the input() function and passes it through
the multiple if statements before placing them in the firstNum string,
the operators object, or the secondNum string.